requirements.txt
alembic==1.2.1
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.4
coveralls==1.8.2
cryptography==2.8
dnspython==1.16.0
docopt==0.6.2
eventlet==0.25.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==0.4.15
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jwt==0.6.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
monotonic==1.5
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
pip-autoremove==0.9.1
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
PySocks==1.7.1
pytest==5.2.0
pytest-cov==2.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.10
tweepy==3.8.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
